refactoring needed to better navigate through the code
