+++ ros/src/tl_detector/light_classification/__init__.py
@@ -1,0 +1,4 @@
+from tl_classifier import TLClassifier
+from opencv_tl_classifier import OpenCVTLClassifier
+from ssd_tl_classifier import SSDTLClassifier
+from yolo_tiny_tl_classifier import YOLOTinyTLClassifier
