+++ ros/src/tl_detector/light_classification/test_tl_classifier.py
@@ -1,6 +1,9 @@
 import rospy
 from unittest import TestCase
-from light_classification.tl_classifier import TLClassifier, OpenCVTrafficLightsClassifier, SSDTrafficLightsClassifier, TinyYOLOTrafficLightsClassifier
+from light_classification.tl_classifier import TLClassifier
+from light_classification.ssd_tl_classifier import SSDTLClassifier
+from light_classification.yolo_tiny_tl_classifier import YOLOTinyTLClassifier
+from light_classification.opencv_tl_classifier import OpenCVTLClassifier
 from roslaunch.parent import ROSLaunchParent
 
 
@@ -23,21 +26,21 @@ class TLClassifierTest(TestCase):
     def setUp(self):
         TLClassifier.INSTANCE = None
         TLClassifier.KNOWN_TRAFFIC_LIGHT_CLASSIFIERS = {}
-        TLClassifier.register_subclass("opencv")(OpenCVTrafficLightsClassifier)
-        TLClassifier.register_subclass("ssd")(SSDTrafficLightsClassifier)
-	TLClassifier.register_subclass("tiny-yolo")(TinyYOLOTrafficLightsClassifier)
+        TLClassifier.register_subclass("opencv")(OpenCVTLClassifier)
+        TLClassifier.register_subclass("ssd")(SSDTLClassifier)
+        TLClassifier.register_subclass("yolo-tiny")(YOLOTinyTLClassifier)
 
     def tearDown(self):
         TLClassifier.INSTANCE = None
         TLClassifier.KNOWN_TRAFFIC_LIGHT_CLASSIFIERS = {}
-        TLClassifier.register_subclass("opencv")(OpenCVTrafficLightsClassifier)
-        TLClassifier.register_subclass("ssd")(SSDTrafficLightsClassifier)
-        TLClassifier.register_subclass("tiny-yolo")(TinyYOLOTrafficLightsClassifier)
+        TLClassifier.register_subclass("opencv")(OpenCVTLClassifier)
+        TLClassifier.register_subclass("ssd")(SSDTLClassifier)
+        TLClassifier.register_subclass("yolo-tiny")(YOLOTinyTLClassifier)
 
     def test_get_instance_of(self):
         instance = TLClassifier.get_instance_of("opencv")
-        self.assertIsInstance(instance, OpenCVTrafficLightsClassifier)
-        self.assertEqual(2, len(TLClassifier.KNOWN_TRAFFIC_LIGHT_CLASSIFIERS))
+        self.assertIsInstance(instance, OpenCVTLClassifier)
+        self.assertEqual(3, len(TLClassifier.KNOWN_TRAFFIC_LIGHT_CLASSIFIERS))
 
     def test_classify(self):
