+++ ros/src/tl_detector/sim_traffic_light_config.yaml
@@ -11,4 +11,4 @@ stop_line_positions:
     - [161.76, 2303.82]
     - [351.84, 1574.65]
 is_site: False
-classifier: "tiny-yolo" # opencv, ssd, simple_cnn, tiny-yolo
+classifier: "yolo-tiny" # opencv, ssd, simple_cnn, yolo-tiny
