+++ ros/src/tl_detector/site_traffic_light_config.yaml
@@ -6,4 +6,4 @@ camera_info:
 stop_line_positions:
     - [8.0, 16.2]
 is_site: True
-classifier: "ssd" # opencv, ssd, simple_cnn, tiny-yolo
+classifier: "ssd" # opencv, ssd, simple_cnn, yolo-tiny
