+++ ros/src/tl_detector/tl_detector.py
@@ -7,7 +7,7 @@ from styx_msgs.msg import Lane
 from sensor_msgs.msg import Image
 from cv_bridge import CvBridge
 from scipy.spatial import KDTree
-from light_classification.tl_classifier import TLClassifier
+from light_classification import TLClassifier
 import tf
 import yaml
